GeoArrays.jl Documentation

GeoArrays.jl Documentation

GeoArrays.jl Documentation

Index

GeoArrays.bbox! — Method.

Set geotransform of GeoArray by specifying a bounding box. Note that this only can result in a non-rotated or skewed GeoArray.

source

GeoArrays.bboxes — Method.

Generate bounding boxes for GeoArray cells.

source

GeoArrays.compose! — Method.

Transform an GeoArray by applying a Transformation.

source

GeoArrays.epsg! — Method.

Set CRS on GeoArray by epsgcode

source

GeoArrays.interpolate! — Method.

Interpolate missing values in GeoArray.

source

GeoArrays.cast_to_gdal — Method.

Converts type of Array for one that exists in GDAL.

source

GeoArrays.centercoords! — Method.

Set AffineMap of GeoArray by specifying the center coordinates for each x, y dimension by a UnitRange.

source

GeoArrays.epsg2wkt — Method.

Get the WKT of an Integer EPSG code

source

GeoArrays.equals — Method.

Check whether two GeoArrayss a and b are geographically equal, although not necessarily in content.

source

GeoArrays.flipud! — Method.

Function to flip GeoArray upside down to adjust to GDAL ecosystem.

source

GeoArrays.get_nodata — Method.

Retrieves nodata value from RasterBand.

source

GeoArrays.is_rotated — Method.

Check wether the AffineMap of a GeoArray contains rotations.

source

GeoArrays.mask_flags — Method.

Takes bitwise OR-ed set of status flags and returns flags.

source

GeoArrays.proj2wkt — Method.

Get the WKT of an Proj string

source

GeoArrays.str2wkt — Method.

Parse CRS string into WKT.

source

Index

GeoArrays.bbox!

GeoArrays.bboxes

GeoArrays.cast_to_gdal

GeoArrays.centercoords!

GeoArrays.compose!

GeoArrays.epsg!

GeoArrays.epsg2wkt

GeoArrays.equals

GeoArrays.flipud!

GeoArrays.get_nodata

GeoArrays.interpolate!

GeoArrays.is_rotated

GeoArrays.mask_flags

GeoArrays.proj2wkt

GeoArrays.str2wkt